services/user-service/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "user-service"
version = "1.0.0"
description = "TalentSync User Service - Authentication and profile management"
authors = [{name = "TalentSync Team", email = "team@talentsync.ai"}]
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "sqlalchemy",
    "alembic",
    "pydantic[email]",
    "passlib[bcrypt]",
    "python-dotenv",
    "python-jose[cryptography]",
    "aiosqlite",
    "python-multipart",
]

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-asyncio",
    "httpx",
    "pytest-cov",
]
dev = [
    "black",
    "isort", 
    "mypy",
    "flake8",
]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.11"
disallow_untyped_defs = true
ignore_missing_imports = true
